Workaround Centos 8 eol installation
CentOS Linux 8 reached End Of Life on December 31st, 2021. The mirror list
with http://mirrorlist.centos.org doesn't work for it. The base URL with
http://mirror.centos.org doesn't work as well. So enabling repos from vault
mirrors:
  https://github.com/CentOS/sig-cloud-instance-images/issues/190

This is not a task of installer.sh, we need to prepare OS before running
installer.sh. installer_centos_8.sh is a script which deos this workaround.
It fixes repositories by enabling repos from vault mirrors then runs
installer.sh as usual.