Fix problem with missing yum on RedOS 7.3
The base Docker image, RedOS 7.3, incorporates the dnf package manager.
However, a minimal image features the microdnf, which is intended to
replace both yum and dnf, albeit with limited functionality. This has
led to an issue where installer.sh does not operate as expected. The
proposed solution rectifies the problem with the operation of
installer.sh on the base image, specifically on the images
registry.red-soft.ru/ubi7/ubi.

Resolves #34